Refactor, R# acceptable names, null tests, kill warnings.

+++ RatTracker-WPF/RatTracker-WPF/APIWorker.cs
@@ -23,10 +23,10 @@ namespace RatTracker_WPF
      */
 	public class ApiWorker
 	{
-		private static readonly ILog logger = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-		private bool stopping;
-		public WebSocket ws;
-		private bool failing;
+		private static readonly ILog Logger = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+		private bool _stopping;
+		public WebSocket Ws;
+		private bool _failing;
 		/*
          * queryAPI sends a GET request to the API. Kindasorta deprecated behavior.
          */
@@ -41,49 +41,49 @@ namespace RatTracker_WPF
 			{
 				const string wsurl = "wss://dev.api.fuelrats.com:443/"; //TODO: Remove this hardcoding! Rewrite URL with ws:// instead of http://
 																   //string wsurl = "ws://dev.api.fuelrats.com/";
-				logger.Info("Connecting to WS at " + wsurl);
-				ws = new WebSocket(wsurl, "", WebSocketVersion.Rfc6455) {AllowUnstrustedCertificate = true};
-				ws.Error += websocketClient_Error;
-				ws.Opened += websocketClient_Opened;
-				ws.MessageReceived += websocketClient_MessageReceieved;
-				ws.Closed += websocket_Client_Closed;
+				Logger.Info("Connecting to WS at " + wsurl);
+				Ws = new WebSocket(wsurl, "", WebSocketVersion.Rfc6455) {AllowUnstrustedCertificate = true};
+				Ws.Error += websocketClient_Error;
+				Ws.Opened += websocketClient_Opened;
+				Ws.MessageReceived += websocketClient_MessageReceieved;
+				Ws.Closed += websocket_Client_Closed;
 				ServicePointManager.ServerCertificateValidationCallback += (sender, cert, chain, sslPolicyErrors) => true;
 			}
 			catch (Exception ex)
 			{
-				logger.Fatal("Well, that went tits up real fast: " + ex.Message);
+				Logger.Fatal("Well, that went tits up real fast: " + ex.Message);
 			}
 		}
 
 		public void OpenWs()
 		{
-			if (ws == null)
+			if (Ws == null)
 			{
-				logger.Debug("Attempted to open uninitialized WS connection.");
+				Logger.Debug("Attempted to open uninitialized WS connection.");
 				return;
 			}
-			ws.Open();
-			logger.Debug("WS client is " + ws.State);
+			Ws.Open();
+			Logger.Debug("WS client is " + Ws.State);
 		}
 
 		public void DisconnectWs()
 		{
-			stopping = true;
-			ws?.Close();
+			_stopping = true;
+			Ws?.Close();
 		}
 
 		public void SendWs(string action, IDictionary<string, string> data)
 		{
-			if (ws == null)
+			if (Ws == null)
 			{
-				logger.Debug("Attempt to send data over uninitialized WS connection!");
+				Logger.Debug("Attempt to send data over uninitialized WS connection!");
 				return;
 			}
 
 			switch (action)
 			{
 				case "rescues:read":
-					logger.Debug("SendWS is requesting a rescues update.");
+					Logger.Debug("SendWS is requesting a rescues update.");
 					break;
 			}
 
@@ -93,65 +93,63 @@ namespace RatTracker_WPF
 				data = data
 			};
 			string json = JsonConvert.SerializeObject(myquery);
-			logger.Debug("sendWS Serialized to: " + json);
-			ws.Send(json);
+			Logger.Debug("sendWS Serialized to: " + json);
+			Ws.Send(json);
 		}
 
 		public void SendQuery(APIQuery myquery)
 		{
 			string json = JsonConvert.SerializeObject(myquery);
-			logger.Debug("Sent an APIQuery serialized as: " + json);
-			ws.Send(json);
+			Logger.Debug("Sent an APIQuery serialized as: " + json);
+			Ws.Send(json);
 		}
 
 		public void SendAuth(AuthQuery myauth)
 		{
 			string json = JsonConvert.SerializeObject(myauth);
-			logger.Debug("Sent an API Auth message serialized as: " + json);
-			ws.Send(json);
+			Logger.Debug("Sent an API Auth message serialized as: " + json);
+			Ws.Send(json);
 		}
 
 		public void SendTpaMessage(TPAMessage message)
 		{
-			if (ws == null)
+			if (Ws == null)
 			{
-				logger.Debug("Attempt to send TPA message over uninitialized WS connection!");
+				Logger.Debug("Attempt to send TPA message over uninitialized WS connection!");
 				return;
 			}
 
 			message.data.Add("platform", "PC");
 			message.applicationId = Properties.Settings.Default.AppID;
 			string json = JsonConvert.SerializeObject(message);
-			logger.Debug("Serialized TPA data: " + json);
-			ws.Send(json);
+			Logger.Debug("Serialized TPA data: " + json);
+			Ws.Send(json);
 		}
 		private async void websocket_Client_Closed(object sender, EventArgs e)
 		{
-			if (stopping)
+			if (_stopping)
 			{
-				logger.Info("Disconnected from API WS server, stopping...");
-				return;
+				Logger.Info("Disconnected from API WS server, stopping...");
 			}
 			else
 			{
-				if(ws.State == WebSocketState.Connecting)
+				if(Ws.State == WebSocketState.Connecting)
 				{
-					logger.Debug("Connection attempt already underway, ignoring reconnect attempt.");
+					Logger.Debug("Connection attempt already underway, ignoring reconnect attempt.");
 					return;
 				}
-				if (failing)
+				if (_failing)
 				{
-					logger.Info("API reconnect still failing. Waiting...");
+					Logger.Info("API reconnect still failing. Waiting...");
 					await Task.Delay(20000);
 					OpenWs();
 					return;
 				}
 
-				logger.Info("API WS Connection closed unexpectedly. Reconnecting in five seconds...");
-				failing = true;
+				Logger.Info("API WS Connection closed unexpectedly. Reconnecting in five seconds...");
+				_failing = true;
 				await Task.Delay(5000);
 				OpenWs();
-				return;
 			}
 		}
 
@@ -164,21 +162,21 @@ namespace RatTracker_WPF
 			//TODO: Implement error handling.
 			if (data.errors != null)
 			{
-				logger.Fatal("API error when deserializing in APIWorker! " + data.data);
+				Logger.Fatal("API error when deserializing in APIWorker! " + data.data);
 				return;
 			}
 
 			//TODO: Implement actual pass to our 3PA logic.
 			if (data.application != null)
 			{
-				logger.Debug("Got an application message, pass to own parser.");
+				Logger.Debug("Got an application message, pass to own parser.");
 				return;
 			}
 
 			switch ((string)data.action)
 			{
 				case "welcome":
-					logger.Info("API MOTD: " + data.data);
+					Logger.Info("API MOTD: " + data.data);
 					break;
 			}
 
@@ -187,15 +185,15 @@ namespace RatTracker_WPF
 
 		public void websocketClient_Error(object sender, ErrorEventArgs e)
 		{
-			logger.Fatal("Websocket: Exception thrown: ", e.Exception);
+			Logger.Fatal("Websocket: Exception thrown: ", e.Exception);
 		}
 
 		public void websocketClient_Opened(object sender, EventArgs e)
 		{
-			failing = false;
-			logger.Info("Websocket: Connection to API established.");
+			_failing = false;
+			Logger.Info("Websocket: Connection to API established.");
 			if (Properties.Settings.Default.OAuthToken == null)
-				logger.Error("Oops! I am attempting to connect without an OAuth token. That's probably bad.");
+				Logger.Error("Oops! I am attempting to connect without an OAuth token. That's probably bad.");
 			else
 			{
 				AuthQuery auth = new AuthQuery()
@@ -211,7 +209,7 @@ namespace RatTracker_WPF
 					data = new Dictionary<string, string> {{"email", Properties.Settings.Default.APIUsername}}
 				};
 				SendQuery(login);
-				logger.Info("Sent login for " + Properties.Settings.Default.APIUsername);
+				Logger.Info("Sent login for " + Properties.Settings.Default.APIUsername);
 			}
 			//TODO: Put stream subscription messages here when Mecha goes live. Do we want to listen to ourselves?
 		}
@@ -220,9 +218,9 @@ namespace RatTracker_WPF
 			Dictionary<string, string> data = new Dictionary<string, string>();
 			data.Add("action", "stream:subscribe");
 			data.Add("applicationId", applicationId);
-			logger.Debug("Subscribing to stream: " + applicationId);
+			Logger.Debug("Subscribing to stream: " + applicationId);
 			string json = JsonConvert.SerializeObject(data);
-			ws.Send(json);
+			Ws.Send(json);
 		}
 		public async Task<string> QueryApi(string action, Dictionary<string, string> data)
 		{
@@ -241,23 +239,23 @@ namespace RatTracker_WPF
 					}
 
 					content.Query = query.ToString();
-					logger.Debug("Built query string:" + content);
+					Logger.Debug("Built query string:" + content);
 					var response = await client.GetAsync(content.ToString());
 					//appendStatus("AsyncPost sent.");
 					if (response.IsSuccessStatusCode)
 					{
 						var responseString = await response.Content.ReadAsStringAsync();
-						logger.Debug("Starting response string parse task.");
+						Logger.Debug("Starting response string parse task.");
 						return ApiGetResponse(responseString);
 					}
 
-					logger.Debug("HTTP request returned an error:" + response.StatusCode);
+					Logger.Debug("HTTP request returned an error:" + response.StatusCode);
 					return "";
 				}
 			}
 			catch (Exception ex)
 			{
-				logger.Fatal("Exception in QueryAPI: ", ex);
+				Logger.Fatal("Exception in QueryAPI: ", ex);
 				return "";
 			}
 		}
@@ -275,15 +273,15 @@ namespace RatTracker_WPF
 
 		public static string ConnectApi()
 		{
-			logger.Info("Oauth token is "+Properties.Settings.Default.OAuthToken);
+			Logger.Info("Oauth token is "+Properties.Settings.Default.OAuthToken);
 			if (Properties.Settings.Default.OAuthToken != "")
 			{
-				logger.Info("Authenticating with OAuth token...");
+				Logger.Info("Authenticating with OAuth token...");
 				return "";
 			}
 
 			/* Connect to the API here. */
-			logger.Debug("No OAuth token stored, attempting to authorize app.");
+			Logger.Debug("No OAuth token stored, attempting to authorize app.");
 			ProcessStartInfo proc = new ProcessStartInfo
 			{
 				UseShellExecute = true,
@@ -291,44 +289,47 @@ namespace RatTracker_WPF
 				FileName = Process.GetCurrentProcess().MainModule.FileName,
 				Verb = "runas"
 			};
-			logger.Debug("Requesting identity.");
+			Logger.Debug("Requesting identity.");
 			WindowsIdentity identity = WindowsIdentity.GetCurrent();
 			WindowsPrincipal principal = new WindowsPrincipal(identity);
-			logger.Debug("Checking registry key...");
+			Logger.Debug("Checking registry key...");
 			if (Microsoft.Win32.Registry.ClassesRoot.OpenSubKey(@"RatTracker\shell\open\command", false) == null)
 			{
-				logger.Debug("No custom URI configured for RatTracker.");
+				Logger.Debug("No custom URI configured for RatTracker.");
 			}
 			else
 			{
-				logger.Debug("An URI handler is set, checking...");
+				Logger.Debug("An URI handler is set, checking...");
 				Microsoft.Win32.RegistryKey key = Microsoft.Win32.Registry.ClassesRoot.OpenSubKey(@"RatTracker\shell\open\command", false);
-				string mypath = (string)key.GetValue("");
-				logger.Debug("Have a URI handler registered: " + mypath + " and my procname is "+proc.FileName);
-				if (!mypath.Contains(proc.FileName))
+				if (key != null)
 				{
-					logger.Debug("URI handler is not set to current version of RatTracker, resetting...");
-				}
-				else
-				{
-					logger.Debug("RatTracker URI is properly configured, launching OAuth authentication process.");
-					var authcontent = new UriBuilder(Path.Combine(Properties.Settings.Default.APIURL + "oauth2/authorise?client_id="+Properties.Settings.Default.ClientID+"&scope=*&redirect_uri=rattracker://auth&state=preinit&response_type=code"))
+					string mypath = (string)key.GetValue("");
+					Logger.Debug("Have a URI handler registered: " + mypath + " and my procname is "+proc.FileName);
+					if (!mypath.Contains(proc.FileName))
 					{
-						Port = Properties.Settings.Default.APIPort
-					};
-					Process.Start(authcontent.ToString());
-					Process.GetCurrentProcess().Kill();
-					return "";
+						Logger.Debug("URI handler is not set to current version of RatTracker, resetting...");
+					}
+					else
+					{
+						Logger.Debug("RatTracker URI is properly configured, launching OAuth authentication process.");
+						var authcontent = new UriBuilder(Path.Combine(Properties.Settings.Default.APIURL + "oauth2/authorise?client_id="+Properties.Settings.Default.ClientID+"&scope=*&redirect_uri=rattracker://auth&state=preinit&response_type=code"))
+						{
+							Port = Properties.Settings.Default.APIPort
+						};
+						Process.Start(authcontent.ToString());
+						Process.GetCurrentProcess().Kill();
+						return "";
+					}
 				}
 			}
 
-			logger.Debug("Passing elevation check.");
+			Logger.Debug("Passing elevation check.");
 			if (!principal.IsInRole(WindowsBuiltInRole.Administrator))
 			{
-				logger.Debug("Not running as administrator, requesting elevation.");
+				Logger.Debug("Not running as administrator, requesting elevation.");
 				try
 				{
-					logger.Debug("Not admin and key does not exist, requesting elevation.");
+					Logger.Debug("Not admin and key does not exist, requesting elevation.");
 					MessageBoxResult result = MessageBox.Show("RatTracker needs to set up a custom URL handler to handle OAuth authentication. This requires elevated privileges. Please click OK to restart RatTracker as Administrator. It will restart with normal privileges once the OAuth process is complete.", "RT Needs elevation", MessageBoxButton.YesNo, MessageBoxImage.Exclamation);
 					switch (result)
 					{
@@ -350,20 +351,29 @@ namespace RatTracker_WPF
 
 			try
 			{
-				logger.Debug("Creating registry key");
+				Logger.Debug("Creating registry key");
 				Microsoft.Win32.RegistryKey key = Microsoft.Win32.Registry.ClassesRoot;
 				key.CreateSubKey("RatTracker", true);
 				key = key.OpenSubKey("RatTracker", true);
-				key.SetValue("URL Protocol", "");
-				key.SetValue("DefaultIcon", "");
-				key.CreateSubKey("shell", true);
-				key = key.OpenSubKey("shell", true);
-				key.CreateSubKey("open", true);
-				key = key.OpenSubKey("open", true);
-				key.CreateSubKey("command", true);
-				key = key.OpenSubKey("command", true);
-				key.SetValue("", "\"" + Process.GetCurrentProcess().MainModule.FileName + "\" \"%1\"");
-				logger.Debug("Launching authorization process.");
+				if (key != null)
+				{
+					key.SetValue("URL Protocol", "");
+					key.SetValue("DefaultIcon", "");
+					key.CreateSubKey("shell", true);
+					key = key.OpenSubKey("shell", true);
+					if (key != null)
+					{
+						key.CreateSubKey("open", true);
+						key = key.OpenSubKey("open", true);
+						if (key != null)
+						{
+							key.CreateSubKey("command", true);
+							key = key.OpenSubKey("command", true);
+							if (key != null) key.SetValue("", "\"" + Process.GetCurrentProcess().MainModule.FileName + "\" \"%1\"");
+						}
+					}
+				}
+				Logger.Debug("Launching authorization process.");
 				var authcontent =
 					new UriBuilder(Path.Combine(Properties.Settings.Default.APIURL +"oauth2/authorise?client_id="+Properties.Settings.Default.ClientID+"&scope=*&redirect_uri=rattracker://auth&state=preinit&response_type=code"))
 					{
@@ -376,7 +386,7 @@ namespace RatTracker_WPF
 			catch (Exception ex)
 			{
 
-				logger.Fatal("Exception in ConnectAPI: ", ex);
+				Logger.Fatal("Exception in ConnectAPI: ", ex);
 				return "";
 			}
 		}
@@ -396,31 +406,31 @@ namespace RatTracker_WPF
          * Primarily used for login, as most of what we need to do is handled through WS.
          * REDONE: Now returns the response as string, process in main.
          */
-        public async Task<string> sendAPI(string action, List<KeyValuePair<string, string>> data)
+        public async Task<string> SendApi(string action, List<KeyValuePair<string, string>> data)
         {
-            logger.Debug("SendAPI was called with action" + action);
+            Logger.Debug("SendAPI was called with action" + action);
             try
             {
                 using (var client = new HttpClient())
                 {
                     var content = new FormUrlEncodedContent(data);
                     var response = await client.PostAsync(Properties.Settings.Default.APIURL + action, content); //TODO: This does not pull port number!
-                    logger.Debug("AsyncPost sent.");
+                    Logger.Debug("AsyncPost sent.");
                     if (response.IsSuccessStatusCode)
                     {
                         var responseString = await response.Content.ReadAsStringAsync();
-                        logger.Debug("Starting response string parse task.");
+                        Logger.Debug("Starting response string parse task.");
                         return ApiResponse(responseString);
                     }
 
-					logger.Info("HTTP request returned an error:" + response.StatusCode);
+					Logger.Info("HTTP request returned an error:" + response.StatusCode);
 	                return "";
 	                //connectWS();
                 }
             }
             catch (Exception ex)
             {
-                logger.Fatal("Well, that didn't go well. SendAPI exception: ", ex);
+                Logger.Fatal("Well, that didn't go well. SendAPI exception: ", ex);
                 return "";
             }
         }
